Force keys to depress when the keyboard hook is disabled to prevent stuck keys when the game gains and loses focus.  Prevented keys being held down from filling the _keysDown list and from firing keydown when the key was already down.
Added GetFriendlyName with ModifierKeys and Keys params for full KeyBinding display strings which could be referenced both by KeyBindings and the new KeybindingAssinger control.

Added the KeybindingAssigner control back into the repo which allows you to tie a KeyBinding to it to then manipulate it.  When double-clicked, a KeybindingAssignmentWindow is shown which allows the user to update the keybinding.

The new controls can be seen in action, here: https://gfycat.com/smoothpreciousbandicoot