Rewrite ArcDps networking code for clarity
Prefer event arguments over same-instance fields for passing event state.

Use a CancellationToken to signal that we want to stop receiving ArcDps messages.

Do not allow SocketListener to dispose itself; raise an event in case of a network error and let the client code decide what to do with the socket.

Skip Shutdown() on sockets that are already disconnected.